menu: Cleanup default items
