+++ src/utils/mod.rs
@@ -5,4 +5,5 @@ pub(crate) use self::ids::id_gen;
 pub mod geometry;
 pub mod iced;
 pub mod prelude;
+pub mod screenshot;
 pub mod tween;
